Island: Move disable_session_cookies to authnetication_service
