+++ monkey/monkey_island/cc/services/authentication_service/configure_flask_security.py
@@ -3,6 +3,8 @@ import secrets
 from pathlib import Path
 from typing import Any, Dict
 
+from flask.sessions import SecureCookieSessionInterface
+
 from common.utils.file_utils import open_new_securely_permissioned_file
 
 SECRET_FILE_NAME = ".flask_security_configuration.json"
@@ -22,3 +24,16 @@ def generate_flask_security_configuration(data_dir: Path) -> Dict[str, Any]:
             json.dump(security_options, secret_file)
 
             return security_options
+
+
+def disable_session_cookies() -> SecureCookieSessionInterface:
+    class CustomSessionInterface(SecureCookieSessionInterface):
+        """Prevent creating session from API requests."""
+
+        def should_set_cookie(self, *args, **kwargs):
+            return False
+
+        def save_session(self, *args, **kwargs):
+            return
+
+    return CustomSessionInterface()
